basic 720p switchover support
new command "720p" causes system to expect a 720p video stream,
overlaying the top left corner of the 1080p raspberry pi stream
onto the 720p stream.

This requires modification to the DMA engine to cleanly genlock
the 720p DMA out over the 1080p input lines.

It also requires some hacks to the video engine setup, which,
as usual, need to be cleaned up once the basic solution is deemed
stable....