staging changes for data island readout so we can change colorspaces
automatic detection of colorspace hinges on being able to read out
infopackets that are in data islands.

These changes create a CSR that captures the most recent data
packet and fires an interrupt, giving the CPU a chance to look
for the infopacket and repsond to it.

Unfortunately, data islands are encrypted, so until we have
authorization to decrypt them, auto-detection and change of
colorspace is too risky to fully implement.